service_peyk/requirements.txt
mysqlclient==1.4.3
Django==4.2
djangorestframework==3.14.0
requests==2.28.2
pymysql==1.0.3